+++ src/slider/Slider.jsx
@@ -1,6 +1,7 @@
 import React from "react";
+import ReactDOM from "react-dom";
 import ShallowComponent from "robe-react-commons/lib/components/ShallowComponent";
-import {Popover, Overlay, Tooltip} from "react-bootstrap";
+import {Popover, Overlay, Tooltip, Col} from "react-bootstrap";
 import "./Slider.css";
 
 export default class Slider extends ShallowComponent {
@@ -15,7 +16,7 @@ export default class Slider extends ShallowComponent {
          */
         minValue: React.PropTypes.number,
         /**
-         * Default value of Slider component (If has range prop value must array like {[0,100]}).
+         * Default value of Slider component (If has range prop defaultValue must array like {[0,100]}).
          */
         defaultValue: React.PropTypes.number,
         /**
@@ -64,24 +65,24 @@ export default class Slider extends ShallowComponent {
         unit: ""
     };
 
-    static idCounter = 1;
+    static counter = 1;
     downEvent = "mousedown";
     upEvent = "mouseup";
     moveEvent = "mousemove";
     click = false;
-    selectedDiv;
-    layerDiv;
-    circleDivMax;
-    circleDivMin;
-    currentId;
+    selectedDivRef;
+    layerDivRef;
+    circleDivMaxRef;
+    circleDivMinRef;
+    currentClassName;
 
     constructor(props:Object) {
         super(props);
-        this.selectedDiv = `SliderSelectedDiv-${Slider.idCounter}`;
-        this.layerDiv = `SliderLayerDiv-${Slider.idCounter}`;
-        this.circleDivMax = `SliderCircleDivMax-${Slider.idCounter}`;
-        this.circleDivMin = `SliderCircleDivMin-${Slider.idCounter}`;
-        Slider.idCounter++;
+        this.selectedDivRef = `SliderSelectedDiv-${Slider.counter}`;
+        this.layerDivRef = `SliderLayerDiv-${Slider.counter}`;
+        this.circleDivMaxRef = `SliderCircleDivMax-${Slider.counter}`;
+        this.circleDivMinRef = `SliderCircleDivMin-${Slider.counter}`;
+        Slider.counter++;
         this.state = {
             valueMax: this.props.range ? this.props.defaultValue ? this.props.defaultValue[1] : this.props.minValue : this.props.defaultValue || this.props.minValue,
             valueMin: this.props.range ? this.props.defaultValue ? this.props.defaultValue[0] : this.props.minValue : this.props.defaultValue || this.props.minValue,
@@ -89,14 +90,15 @@ export default class Slider extends ShallowComponent {
             defaultMinPx: 0,
             openMaxDesc: false,
             openMinDesc: false,
-            disabled: this.props.disabled
+            disabled: this.props.disabled,
+            unit: this.props.unit
         };
     }
 
     render():Object {
 
         let classNameSelected = !this.state.disabled ? "sliderSelectedArea" : "sliderSelectedAreaDisabled";
-        let classNameCircle = !this.state.disabled ? "sliderButton" : "sliderButtonDisabled";
+        let classNameCircle = !this.state.disabled ? "sliderButtonMax" : "sliderButtonMaxDisabled";
         let classNameCircleMin = !this.state.disabled ? "sliderButtonMin" : "sliderButtonMinDisabled";
         let classNameLayer = !this.state.disabled ? "sliderLayer" : "sliderLayerDisabled";
 
@@ -106,35 +108,31 @@ export default class Slider extends ShallowComponent {
         return (<span>
             <Overlay show={this.state.openMaxDesc}
                      placement="top"
-                     container={document.getElementById(this.circleDivMax)}
-                     target={document.getElementById(this.circleDivMax)}>
-          <Tooltip id="sliderPopover">{this.state.valueMax + " " + this.props.unit}</Tooltip>
+                     container={this.circleMaxDOM}
+                     target={this.circleMaxDOM}>
+          <Tooltip id="sliderPopover">{this.state.valueMax + " " + this.state.unit}</Tooltip>
         </Overlay>
             <Overlay show={this.state.openMinDesc}
                      placement="top"
-                     container={document.getElementById(this.circleDivMin)}
-                     target={document.getElementById(this.circleDivMin)}>
-          <Tooltip id="sliderPopover">{this.state.valueMin + " " + this.props.unit}</Tooltip>
+                     container={this.circleMinDOM}
+                     target={this.circleMinDOM}>
+          <Tooltip id="sliderPopover">{this.state.valueMin + " " + this.state.unit}</Tooltip>
         </Overlay>
-            <div className="sliderContainer">
-            <div id={this.selectedDiv} className={classNameSelected}
+            <Col className="sliderContainer">
+            <Col ref={this.selectedDivRef} className={classNameSelected}
                  style={{left: styleLeft, width: styleWidth}}
-                 onClick={!this.state.disabled ? this.__layerClick : null}></div>
-            <div id={this.layerDiv} className={classNameLayer}
-                 onClick={!this.state.disabled ? this.__layerClick : null}></div>
-            <div id={this.circleDivMax} className={classNameCircle}
+                 onClick={!this.state.disabled ? this.__layerClick : null}/>
+            <Col ref={this.layerDivRef} className={classNameLayer}
+                 onClick={!this.state.disabled ? this.__layerClick : null}/>
+            <Col ref={this.circleDivMaxRef} className={classNameCircle}
                  style={{left: parseFloat(this.state.defaultMaxPx)}}
-                 onMouseOver={()=> this.setState({openMaxDesc: true, openMinDesc: false})}
-                 onMouseDown={()=> this.setState({openMaxDesc: true, openMinDesc: false})}
-                 onMouseUp={()=> this.setState({openMaxDesc: false, openMinDesc: false})}>
-            </div>
+                 onMouseOver={()=> this.setState({openMaxDesc: true, openMinDesc: false})}>
+            </Col>
                 {this.props.range ?
-                    <div id={this.circleDivMin} className={classNameCircleMin}
+                    <Col ref={this.circleDivMinRef} className={classNameCircleMin}
                          style={{left: this.state.defaultMinPx}}
-                         onMouseOver={()=> this.setState({openMinDesc: true, openMaxDesc: false})}
-                         onMouseDown={()=> this.setState({openMinDesc: true, openMaxDesc: false})}
-                         onMouseUp={()=> this.setState({openMinDesc: false, openMaxDesc: false})}>
-                    </div> : null}
+                         onMouseOver={()=> this.setState({openMinDesc: true, openMaxDesc: false})}>
+                    </Col> : null}
                 {!this.props.closeLabel ?
                     <span>
                         <span
@@ -142,20 +140,20 @@ export default class Slider extends ShallowComponent {
                         <span
                             className="pull-right sliderLabel">{this.props.maxLabel ? this.props.maxLabel : this.props.maxValue}</span>
                     </span> : null}
-        </div>
+        </Col>
         </span>);
     }
 
     __calculateStyles(e:Object) {
-        var selected = document.getElementById(this.selectedDiv);
-        var circle = document.getElementById(this.circleDivMax);
-        var layer = document.getElementById(this.layerDiv);
+        var selected = this.selectedDOM;
+        var circle = this.circleMaxDOM;
+        var layer = this.layerDOM;
         let pageX = this.moveEvent === "mousemove" ? e.clientX : this.click ? e.clientX : e.touches[0].pageX;
         let calculatedPosition = parseInt(pageX - selected.getBoundingClientRect().left);
 
         if (!this.props.step) {
             if (calculatedPosition >= 0 && calculatedPosition <= layer.offsetWidth)
-                this.__changeStyles(calculatedPosition, selected, circle, layer);
+                this.__changeStyles(e, calculatedPosition, selected, circle, layer);
         }
         else {
             if (calculatedPosition >= 0 && calculatedPosition <= layer.offsetWidth) {
@@ -165,9 +163,9 @@ export default class Slider extends ShallowComponent {
                     let halfOfStep = i + halfOfSteps;
                     if (calculatedPosition >= i && calculatedPosition <= i + steps) {
                         if (calculatedPosition < halfOfStep)
-                            this.__changeStyles(i, selected, circle, layer);
+                            this.__changeStyles(e, i, selected, circle, layer);
                         else if (calculatedPosition > halfOfStep)
-                            this.__changeStyles(i + steps, selected, circle, layer);
+                            this.__changeStyles(e, i + steps, selected, circle, layer);
                     }
                 }
             }
@@ -175,17 +173,17 @@ export default class Slider extends ShallowComponent {
     };
 
     __calculateStylesWithRange(e:Object) {
-        var selected = document.getElementById(this.selectedDiv);
-        var circleMax = document.getElementById(this.circleDivMax);
-        var circleMin = document.getElementById(this.circleDivMin);
-        var layer = document.getElementById(this.layerDiv);
+        var selected = this.selectedDOM;
+        var circleMax = this.circleMaxDOM;
+        var circleMin = this.circleMinDOM;
+        var layer = this.layerDOM;
         let pageX = this.moveEvent === "mousemove" ? e.clientX : this.click ? e.clientX : e.touches[0].pageX;
         let calculatedPosition = parseInt(pageX - layer.getBoundingClientRect().left);
 
         if (!this.props.step) {
             if (calculatedPosition >= 0 && calculatedPosition <= layer.offsetWidth) {
 
-                if (this.currentId.startsWith("SliderCircleDivMin")) {
+                if (this.currentClassName.startsWith("sliderButtonMin")) {
                     circleMin.style.left = calculatedPosition + 'px';
                     if (parseInt(circleMin.style.left) >= parseInt(circleMax.style.left)) {
                         selected.style.left = parseFloat(circleMax.style.left) + 'px';
@@ -200,15 +198,21 @@ export default class Slider extends ShallowComponent {
                     let lastValue = this.props.step && this.__isFloat(this.props.step) ? parseFloat(this.props.minValue + newValue).toFixed((this.props.step).toString().split(".")[1].length) : Math.round(this.props.minValue + newValue);
 
                     if (this.props.onChange && lastValue !== this.state.valueMin) {
-                        if (lastValue > this.state.valueMax)
-                            this.props.onChange([this.state.valueMax, lastValue]);
-                        else
-                            this.props.onChange([lastValue, this.state.valueMax]);
+                        if (lastValue > this.state.valueMax) {
+                            e.target.value = [this.state.valueMax, lastValue];
+                            e.target.parsedValue = [this.state.valueMax, lastValue];
+                            this.props.onChange(e);
+                        }
+                        else {
+                            e.target.value = [lastValue, this.state.valueMax];
+                            e.target.parsedValue = [lastValue, this.state.valueMax];
+                            this.props.onChange(e);
+                        }
                     }
 
                     this.setState({valueMin: lastValue});
                 }
-                if (this.currentId.startsWith("SliderCircleDivMax")) {
+                if (this.currentClassName.startsWith("sliderButtonMax")) {
                     circleMax.style.left = (calculatedPosition) + 'px';
                     if (parseInt(circleMin.style.left) <= parseInt(circleMax.style.left)) {
                         selected.style.left = parseFloat(circleMin.style.left) + 'px';
@@ -223,10 +227,16 @@ export default class Slider extends ShallowComponent {
                     let lastValue = this.props.step && this.__isFloat(this.props.step) ? parseFloat(this.props.minValue + newValue).toFixed((this.props.step).toString().split(".")[1].length) : Math.round(this.props.minValue + newValue);
 
                     if (this.props.onChange && lastValue !== this.state.valueMax) {
-                        if (lastValue < this.state.valueMin)
-                            this.props.onChange([lastValue, this.state.valueMin]);
-                        else
-                            this.props.onChange([this.state.valueMin, lastValue]);
+                        if (lastValue < this.state.valueMin) {
+                            e.target.value = [lastValue, this.state.valueMin];
+                            e.target.parsedValue = [lastValue, this.state.valueMin];
+                            this.props.onChange(e);
+                        }
+                        else {
+                            e.target.value = [this.state.valueMin, lastValue];
+                            e.target.parsedValue = [this.state.valueMin, lastValue];
+                            this.props.onChange(e);
+                        }
                     }
 
                     this.setState({valueMax: lastValue});
@@ -239,15 +249,15 @@ export default class Slider extends ShallowComponent {
                 let halfOfSteps = steps / 2;
                 for (let i = 0; i <= parseInt(layer.offsetWidth); i += steps) {
                     let halfOfStep = i + halfOfSteps;
-                    if (this.currentId.startsWith("SliderCircleDivMin")) {
+                    if (this.currentClassName.startsWith("sliderButtonMin")) {
                         if (calculatedPosition >= i && calculatedPosition <= i + steps) {
                             if (calculatedPosition < halfOfStep) {
                                 circleMin.style.left = i + 'px';
-                                this.__changeStylesWithRange(selected, layer, circleMin, circleMax, i, "min");
+                                this.__changeStylesWithRange(e, selected, layer, circleMin, circleMax, i, "min");
                             }
                             else if (calculatedPosition > halfOfStep) {
                                 circleMin.style.left = (i + steps) + 'px';
-                                this.__changeStylesWithRange(selected, layer, circleMin, circleMax, i + steps, "min");
+                                this.__changeStylesWithRange(e, selected, layer, circleMin, circleMax, i + steps, "min");
                             }
 
                             if (parseInt(circleMin.style.left) >= parseInt(circleMax.style.left)) {
@@ -256,15 +266,15 @@ export default class Slider extends ShallowComponent {
                             }
                         }
                     }
-                    if (this.currentId.startsWith("SliderCircleDivMax")) {
+                    if (this.currentClassName.startsWith("sliderButtonMax")) {
                         if (calculatedPosition >= i && calculatedPosition <= i + steps) {
                             if (calculatedPosition < halfOfStep) {
                                 circleMax.style.left = i + 'px';
-                                this.__changeStylesWithRange(selected, layer, circleMin, circleMax, i, "max");
+                                this.__changeStylesWithRange(e, selected, layer, circleMin, circleMax, i, "max");
                             }
                             else if (calculatedPosition > halfOfStep) {
                                 circleMax.style.left = (i + steps) + 'px';
-                                this.__changeStylesWithRange(selected, layer, circleMin, circleMax, i + steps, "max");
+                                this.__changeStylesWithRange(e, selected, layer, circleMin, circleMax, i + steps, "max");
                             }
 
                             if (parseInt(circleMin.style.left) >= parseInt(circleMax.style.left)) {
@@ -278,7 +288,7 @@ export default class Slider extends ShallowComponent {
         }
     };
 
-    __changeStylesWithRange(selected, layer, circleMin, circleMax, step, state) {
+    __changeStylesWithRange(e, selected, layer, circleMin, circleMax, step, state) {
         let decideState = state === "min" ? "valueMin" : "valueMax";
         selected.style.width = (parseFloat(circleMax.style.left) - parseFloat(circleMin.style.left)) + 'px';
         selected.style.left = parseFloat(circleMin.style.left) + 'px';
@@ -289,24 +299,33 @@ export default class Slider extends ShallowComponent {
         let decideMin = decideState === "valueMin" ? lastValue : this.state.valueMin;
         let decideMax = decideState === "valueMax" ? lastValue : this.state.valueMax;
         if (this.props.onChange && decideMin !== this.state.valueMin || this.props.onChange && decideMax !== this.state.valueMax) {
-            if (decideMin > decideMax)
-                this.props.onChange([decideMax, decideMin]);
-            else
-                this.props.onChange([decideMin, decideMax]);
+            if (decideMin > decideMax) {
+                e.target.value = [decideMax, decideMin];
+                e.target.parsedValue = [decideMax, decideMin];
+                this.props.onChange(e);
+            }
+            else {
+                e.target.value = [decideMin, decideMax];
+                e.target.parsedValue = [decideMin, decideMax];
+                this.props.onChange(e);
+            }
         }
 
         this.setState({[decideState]: lastValue});
     };
 
-    __changeStyles(calculatedPosition, selected, circle, layer) {
+    __changeStyles(e, calculatedPosition, selected, circle, layer) {
         selected.style.width = calculatedPosition + 'px';
         circle.style.left = calculatedPosition + 'px';
         let calculateValue = layer.offsetWidth / (this.props.maxValue - this.props.minValue);
         let newValue = calculatedPosition / calculateValue;
         let lastValue = this.props.step && this.__isFloat(this.props.step) ? parseFloat(this.props.minValue + newValue).toFixed((this.props.step).toString().split(".")[1].length) : Math.round(this.props.minValue + newValue);
 
-        if (this.props.onChange && this.state.valueMax != lastValue)
-            this.props.onChange(lastValue);
+        if (this.props.onChange && this.state.valueMax != lastValue) {
+            e.target.value = lastValue;
+            e.target.parsedValue = lastValue;
+            this.props.onChange(e);
+        }
 
         this.setState({valueMax: lastValue});
     };
@@ -314,7 +333,7 @@ export default class Slider extends ShallowComponent {
     __calculateDefault(type:string) {
         let range = type === "min" ? 0 : 1;
         let defaultValue = this.props.range ? this.props.defaultValue ? this.props.defaultValue[range] : this.props.minValue : this.props.defaultValue ? this.props.defaultValue : (this.props.minValue || 0);
-        var layer = document.getElementById(this.layerDiv);
+        var layer = this.layerDOM;
         if (layer !== undefined && layer !== null) {
             let calculatePosition = Math.round(parseFloat((((layer.offsetWidth * defaultValue) / (this.props.maxValue - this.props.minValue)) - ((layer.offsetWidth / (this.props.maxValue - this.props.minValue)) * this.props.minValue)).toFixed(2)));
             return calculatePosition + "px";
@@ -343,25 +362,54 @@ export default class Slider extends ShallowComponent {
     }
 
     __mouseUp() {
+        if (!this.props.range && this.refs[this.circleDivMaxRef])
+            this.setState({openMaxDesc: false});
+        else if (this.props.range && this.refs[this.circleDivMinRef])
+            this.setState({openMinDesc: false, openMaxDesc: false});
+
         document.removeEventListener(this.moveEvent, this.__circleDivMove, true);
     };
 
     __mouseDown(e:Object) {
         e.preventDefault();
+        this.currentClassName = e.target.className;
+
+        if (!this.props.range && this.refs[this.circleDivMaxRef])
+            this.setState({openMaxDesc: true});
+        else if (this.props.range && this.refs[this.circleDivMinRef]) {
+            if (this.currentClassName.startsWith("sliderButtonMin"))
+                this.setState({openMinDesc: true, openMaxDesc: false});
+            else
+                this.setState({openMinDesc: false, openMaxDesc: true});
+        }
+
         document.addEventListener(this.moveEvent, this.__circleDivMove, true);
-        this.currentId = e.target.id;
     };
 
     __resize() {
+        if (!this.props.range && this.refs[this.circleDivMaxRef] && this.refs[this.selectedDivRef] && this.refs[this.layerDivRef])
+            this.setState({
+                defaultMaxPx: this.__calculateDefault("max")
+            });
+        else if (this.props.range && this.refs[this.circleDivMaxRef] && this.refs[this.circleDivMinRef] && this.refs[this.selectedDivRef] && this.refs[this.layerDivRef])
+            this.setState({
+                defaultMaxPx: this.__calculateDefault("max"),
+                defaultMinPx: this.__calculateDefault("min")
+            });
+    };
+
+    componentWillReceiveProps(nextProps) {
         this.setState({
-            defaultMaxPx: this.__calculateDefault("max"),
-            defaultMinPx: this.__calculateDefault("min")
+            disabled: nextProps.disabled,
+            unit: nextProps.unit
         });
     };
 
     componentDidMount() {
-        let circleDivMax = document.getElementById(this.circleDivMax);
-        let circleDivMin = document.getElementById(this.circleDivMin);
+        this.circleMaxDOM = ReactDOM.findDOMNode(this.refs[this.circleDivMaxRef]);
+        this.circleMinDOM = ReactDOM.findDOMNode(this.refs[this.circleDivMinRef]);
+        this.selectedDOM = ReactDOM.findDOMNode(this.refs[this.selectedDivRef]);
+        this.layerDOM = ReactDOM.findDOMNode(this.refs[this.layerDivRef]);
 
         if (/Android|webOS|iPhone|iPad|iPod|BlackBerry|IEMobile|Opera Mini/i.test(navigator.userAgent)) {
             this.downEvent = "touchstart";
@@ -370,10 +418,10 @@ export default class Slider extends ShallowComponent {
         }
 
         if (!this.state.disabled) {
-            circleDivMax.addEventListener(this.downEvent, this.__mouseDown, false);
+            this.circleMaxDOM.addEventListener(this.downEvent, this.__mouseDown, false);
             if (this.props.range) {
-                circleDivMin.addEventListener(this.downEvent, this.__mouseDown, false);
-                this.currentId = this.circleDivMin;
+                this.circleMinDOM.addEventListener(this.downEvent, this.__mouseDown, false);
+                this.currentClassName = this.circleMinDOM.className;
             }
             document.addEventListener(this.upEvent, this.__mouseUp, false);
         }
@@ -386,10 +434,7 @@ export default class Slider extends ShallowComponent {
         });
     };
 
-    componentWillReceiveProps(nextProps) {
-        this.setState({
-            disabled: nextProps.disabled
-        });
+    componentWillUnmount(){
+        window.removeEventListener("resize", this.__resize, true);
     };
-
 }